requirements.txt
discord.py
asyncio
SQLAlchemy
PyGithub
